requirements.txt
pandas
python-bcb
pyyaml
sqlalchemy

